GooglePayReconfigure: Move drop-in view models to their own package
