+++ drop-in/src/main/java/com/adyen/checkout/dropin/ui/stored/PreselectedStoredPaymentMethodFragment.kt
@@ -34,11 +34,12 @@ import com.adyen.checkout.dropin.getComponentFor
 import com.adyen.checkout.dropin.ui.base.DropInBottomSheetDialogFragment
 import com.adyen.checkout.dropin.ui.paymentmethods.GenericStoredModel
 import com.adyen.checkout.dropin.ui.paymentmethods.StoredCardModel
-import com.adyen.checkout.dropin.ui.stored.PreselectedStoredState.AwaitingComponentInitialization
-import com.adyen.checkout.dropin.ui.stored.PreselectedStoredState.PaymentError
-import com.adyen.checkout.dropin.ui.stored.PreselectedStoredState.RequestPayment
-import com.adyen.checkout.dropin.ui.stored.PreselectedStoredState.ShowStoredPaymentDialog
+import com.adyen.checkout.dropin.ui.viewmodel.PreselectedStoredState.AwaitingComponentInitialization
+import com.adyen.checkout.dropin.ui.viewmodel.PreselectedStoredState.PaymentError
+import com.adyen.checkout.dropin.ui.viewmodel.PreselectedStoredState.RequestPayment
+import com.adyen.checkout.dropin.ui.viewmodel.PreselectedStoredState.ShowStoredPaymentDialog
 import com.adyen.checkout.dropin.ui.viewModelsFactory
+import com.adyen.checkout.dropin.ui.viewmodel.PreselectedStoredPaymentViewModel
 
 private val TAG = LogUtil.getTag()
 private const val STORED_PAYMENT_KEY = "STORED_PAYMENT"
