+++ drop-in/src/main/java/com/adyen/checkout/dropin/ui/viewmodel/DropInActivityEvent.kt
@@ -6,7 +6,7 @@
  * Created by josephj on 29/11/2021.
  */
 
-package com.adyen.checkout.dropin.ui
+package com.adyen.checkout.dropin.ui.viewmodel
 
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.model.payments.request.OrderRequest
